cleanup: converge duplicated deploy/fragment primitives, drop stale hooks
- scripts/deploy: single owner motion_primitives.py for the 6 verbatim
  warmup/cooldown functions (bin IO, FixStand FK, hermite, slerp, ang vel)
- base/scene.py: single owner resolve_scene_fragment_path; drop verbatim
  copies in mujoco/xml.py and motrix/scene.py
- drop zero-reference stale hooks: appo eval_mode/get_weights, him_ppo
  test_mode/add_transitions, fast_td3 soft_update alias, ExceptionWrapper.reraise
- drop unreachable return in benchmark_mlp_inference._safe_envs_per_sec
