+++ src/unilab/algos/torch/him_ppo/algorithm.py
@@ -75,9 +75,6 @@ class HIMPPO:
             self.device,
         )
 
-    def test_mode(self) -> None:
-        self.actor_critic.eval()
-
     def train_mode(self) -> None:
         self.actor_critic.train()
